ObeliskFarm/requirements.txt
# Python-Abhängigkeiten für das Test-Skript
# Installieren mit: pip install -r requirements.txt

Pillow>=10.0.0
numpy>=1.24.0

# Optional, aber empfohlen für erweiterte Bildverarbeitung:
opencv-python>=4.8.0  # Für Icon-Erkennung und Template Matching
easyocr>=1.7.0        # Für OCR (Text-Erkennung) - KEINE Systeminstallation nötig!
pytesseract>=0.3.10   # Alternative OCR (nur wenn benötigt)
                     # ACHTUNG: Tesseract muss zusätzlich installiert werden:
                     # Windows: https://github.com/UB-Mannheim/tesseract/wiki
                     # Mac: brew install tesseract
                     # Linux: sudo apt-get install tesseract-ocr

